Change to question set
Towards improving precision and recall of templeKB, we changed the question model. The question set was changed from what is presented in paper to one in Questions.py .  We regenerated the data as templeKB 2.0. It has the following improvements over templeKB 1.0

Location	Deity	Age	Festival	Key-people	Trivia	Manage	Language	Address
templeKB (paper)	3.211169284	2.694589878	1.338568935	0.942408377	0.0610820244	0.3612565445	0	0	0
templeKB2 (new)	1.968152866	1.388535032	1.312101911	0.949044586	0.8343949045	1.159235669	0.4522292994	0.3503184713	0.949044586
Improvement (in %)	-63.15649763	-94.05991322	-2.01714702	0.6992515549	92.67948257	68.83666072	100	100	100

We have good improvement in management, Language and Address. But recall on location and deity has gone down. We are looking at ways of improving this with static dictionaries. 

Also note that 'Festivals' and 'Trivia' are new headings.